Move: redis stomp 설정 파일 이동
